pyproject.toml
[project]
name = "gmailstream"
version = "0.2.2"
description = "CLI tool to download Gmail messages matching configurable filters via OAuth2"
authors = [{ name = "Tiago Silva" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
keywords = ["gmail", "email", "download", "oauth2", "cli", "google", "attachments", "backup"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Email",
    "Environment :: Console",
    "Operating System :: OS Independent",
]
dependencies = [
    "google-api-python-client",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "click",
    "pyyaml",
]

[project.scripts]
gmailstream = "gmailstream.cli:main"

[project.urls]
Homepage = "https://github.com/tsilva/gmailstream"
Repository = "https://github.com/tsilva/gmailstream"
Issues = "https://github.com/tsilva/gmailstream/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gmailstream"]

[tool.hatch.build.hooks.custom]
path = "hatch_build.py"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
